:recycle: Smart 로직 없는 공통 컴포넌트 container 제거
