+++ src/components/SearchHistoryButton/SearchHistoryButton.jsx
@@ -1,18 +1,39 @@
-import React, { useState } from "react";
+import React from "react";
 
-import SearchHistoryButtonContainer from "./SearchHistoryButtonContainer";
+import styled from "styled-components";
 
-const SearchHistoryButton = () => {
-  const [isClicked, setIsClicked] = useState(false);
-  const handleClickButton = () => {
-    setIsClicked((prev) => !prev);
-  };
+import { colors, SButton, SText } from "@styles";
+import { lightgreen } from "@styles/block.style";
+import { g1, green } from "@styles/text.style";
+
+const SearchHistoryButton = ({ isClicked = false, onClick }) => {
   return (
-    <SearchHistoryButtonContainer
+    <S.Button
+      col
+      ct
+      br={12}
+      w={81}
+      h={40}
+      onClick={onClick}
       isClicked={isClicked}
-      onClick={handleClickButton}
-    />
+    >
+      <S.ButtonText b3 isClicked={isClicked}>
+        검색 기록
+      </S.ButtonText>
+    </S.Button>
   );
 };
+
+const S = {};
+
+S.ButtonText = styled(SText)`
+  ${(props) => (props.isClicked ? green : g1)}
+`;
+
+S.Button = styled(SButton)`
+  border: 1px solid
+    ${(props) => (props.isClicked ? colors.green : colors.gray3)};
+  ${(props) => props.isClicked && lightgreen}
+`;
 
 export default SearchHistoryButton;
